+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## HEAD
 
+- Rescue 403 on pipeline delete (https://github.com/heroku/hatchet/pull/130)
 - Performance improvement when running without an explicit HEROKU_API_KEY set (https://github.com/heroku/hatchet/pull/128)
 - Additional rate throttle cases handled (https://github.com/heroku/hatchet/pull/128)